business_logic/requirements.txt
Flask==2.3.2
requests==2.31.0
python-dateutil==2.8.2
prometheus_client==0.17.1